UPSTREAM_OR_FROMLIST: CHROMIUM: MALI: Add support for Rockchip
This is a copy of what we had in the Chrome OS 4.4 kernel as of commit
50549bb894e8 ("UPSTREAM: HID: intel-ish-hid: Use VID/PID from ISH")
but with the following changes:

We handle the fact that CL:1325749 changed the mali platform code to
expect an array of regulators instead of a single regulator.

We handle the fact that CL:1433153 caused the mali platform code to no
longer prepare the clock for us.  ...so we switch to use a
prepare_enable, which seems to work (though I haven't done any deep
digging to check clock enables/disables).

We set PLATFORM_POWER_DOWN_ONLY.  As of CL:1539807 I have extended the
mali platform code to allow defining this for a given platform.  NOTE:
obviously this wouldn't be an upstream solution since upstream would
require things like this to be decided at runtime so you could have a
single kernel for multiple devices, but with the current setup of the
mali platform code there isn't tons of choice here.

BUG=chromium:941638
TEST=With device tree + config the GPU works
TEST=git diff cros/chromeos-4.4 drivers/*/*/midgard/*/rk matches desc.

Change-Id: Ife3bb7175b78158e184629753acba2b52945d0db
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1539808
Reviewed-by: Dominik Behr <dbehr@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>