CHROMIUM: MALI: Fix printf code for 'seqno'
As of CL:1704173, seqno is 64-bit.  We need to adjust our printf
strings to account for this.  Otherwise, we get warnings like:

.../mali_kbase_jd_debugfs.c:
  In function 'kbase_jd_debugfs_fence_info':
.../mali_kbase_jd_debugfs.c:86:7:
  error: format '%u' expects argument of type 'unsigned int',
  but argument 4 has type 'u64' [-Werror=format=]

Technically seqno moved to 64-bit in kernel 5.1, but we've backported
it to our 4.19 kernel so we can't exactly add a special-version check.

BUG=chromium:941638
TEST=Build and boot

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1709132
Reviewed-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit c561f7c4b869e10e0ce02aa03c52eb2707adf702)
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>

Change-Id: I89e2cbe2e7d6bec892206a5c647729e48d2645d4