BL_INTERNAL:drm/panel: Add pumpkin display support
This driver is a copy of panel-raspberrypi-touchscreen.c modified to
support pumpkin.

This is not meant for upstream: this is a separate driver in order
to not break the raspberrypi driver since I don't have time to make
and test on both board a driver that would support both at the same
time.

Note: This is some issue in the horizontal frontporch but not big
enough for me to pass time on it yet.

Signed-off-by: Fabien Parent <fparent@baylibre.com>