BL_INTERNAL:Bluetooth: btmtksdio: fix close() pm for HCI_CHANNEL_USER
When calling close() on a HCI_CHANNEL_USER socket, the vendor specific
shutdown() function is not called.

This example reproduces the issue:

    struct sockaddr_hci addr;
    int ctl;

    // error handling omitted
    ctl = socket(AF_BLUETOOTH, SOCK_RAW, BTPROTO_HCI));

    memset(&addr, 0, sizeof(addr));

    addr.hci_family = AF_BLUETOOTH;
    addr.hci_dev = 0;
    addr.hci_channel = HCI_CHANNEL_USER;

    // error handling omitted
    bind(ctl, (struct sockaddr*)&addr, sizeof(addr));

    close(ctl); // btmtksdio_{flush,close}() are called, not shutdown()

In that case, we never call pm_runtime_{put_noidle,disable}(), which
makes the pm subsystem believe that the device is still being used.

Fix it by moving both calls to btmtksdio_close(), which is always called.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>