ASoC: Fix the codec names used by the Vesper soundcard
The index in the codec names used by the Vesper soundcard have
changed. As a consequence, fixing them accordingly to have the
soundcard working correctly. So that it does not happen again
the hardcoded names should be avoided or at least we should make
sure they do not change as described in MRI-71.

Signed-off-by: Nicolas Belin <nbelin@baylibre.com>