media: mtk-vpu: add vcu driver
The VCU driver is used to communicate with a userspace daemon instead
of a co-processor. I ported the MediaTek VCU driver but didn't clean
it up because it will take a while to make it in an upstreamable way.
The clean-up will happen at a later time during the upstreaming.

Signed-off-by: Fabien Parent <fparent@baylibre.com>