BL_INTERNAL:remoteproc: Check if has_iommu is set before to release the iommu
remoteproc try to use iommu_domain_alloc() which doesn't work on
MT8183. Still, we have to configure the IOMMU for this platform.
Even if we can't allocate a domain, we can get one from the platform
driver using iommu_get_domain_for_dev().
But if we set a domain from the platform driver, then the remoteproc
driver may try to release it which will cause a crash or break the
other drivers using the IOMMU.
To avoid that, check if has_iommu is set (required to allocate the
domain) before to try to release the domain.

Note:
We have tried to update the IOMMU driver to support
iommu_domain_alloc() but without success so far.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>