BL_INTERNAL:ASoC: codecs: limitate the number of channels to the number of TDM slots
Limitate the number of channels that the ADC is outputting so that
it fits in the number of allocated slots on the TDM. This is
useful when more than one codec is used and each codec has its
own allocated slots. In this case the number of channels in the
hw_params function in the total number of channels and the number
of channels given by set_tdm_slot is the real number of channels
to output for this specific codec.

Signed-off-by: Nicolas Belin <nbelin@baylibre.com>