UPSTREAM_OR_FROMLIST: devfreq: add mediatek cci devfreq
This adds a devfreq driver for the Cache Coherent Interconnect (CCI)
of the Mediatek MT8183.

On the MT8183 the CCI is supplied by the same regulator as the LITTLE
cores. The driver is notified when the regulator voltage changes
(driven by cpufreq) and adjusts the CCI frequency to the maximum
possible value.

Change-Id: I3d54ede336418e76df5d316064b6c3857a9f1075
Signed-off-by: Andrew-sh.Cheng <andrew-sh.cheng@mediatek.com>