ARM64: dts: mediatek: rename pumpkin dtsi
Until now there was two types of pumpkin boards and both were based
on compatible SoC: MT8167/MT8516.
A new pumpkin board is being developed and will include a MT8183 SoC.

Rename the pumpkin-common.dtsi device-tree to refer to which platform
it correspond to: pumpkin-common.dtsi -> i300-pumpkin-common.dtsi.

i300 is a platform that includes MT8516/MT8167 as SoC and MT6392 for
the PMIC.

Signed-off-by: Fabien Parent <fparent@baylibre.com>