media: mtk-mdp: assign colorspace also for output
The colorspace needs to also be defined for the output. Currently
the colorspace for the output will be by default the same one
that was for capture since the struct is reused and never cleared.
In case the colorspace for the output is not similar to the one
for the capture, then userspace will get a wrong colorspace.

Signed-off-by: Fabien Parent <fparent@baylibre.com>