BL_INTERNAL:Input: mtk-kpd - initialize keymap_state to match hardware
MediaTek's keypad key status are reported in the KPD_MEM1 -> KPD_MEM5
registers as following:
- 1: key is released (default)
- 0: key is pressed

We only use the 16 lower bits of each register (KPD_MEM{1..5})

According to regmap, these registers default as following:

$ cat /sys/kernel/debug/regmap/10010000.kp/registers
  00: 00000000 (KPD_STA)
  04: 0000ffff (KPD_MEM1)
  08: 0000ffff (KPD_MEM2)
  0c: 0000ffff (KPD_MEM3)
  10: 0000ffff (KPD_MEM4)
  14: 0000ffff (KPD_MEM5)

However, the keymap_state bitmap is initialized to:
  ff,ffffffff,ffffffff,ffffffff,ffffffff

This is incorrect. Because of this, the driver reports lots of bogus
events the first time we use the keypad.

Initialize the keymap_state to:
  ff,0000ffff,0000ffff,0000ffff,0000ffff

Which fixes the above issue.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>