rpmsg: apu_rpmsg: Update the way to get a remoteproc device
Currently, we use the device tree to get a remoteproc device.
The way this is done is quite hacky and is not relliable and
scalable, neither upstreamable.
This updates the way to get a remoteproc device.

Note:
currently, this only work when the parent device is a virtio
device. Another type of parent is not supported.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>