UPSTREAM_OR_FROMLIST: media: platform: mtk-isp: camsv: Locate seninf pad dynamically
When creating the link between the seninf and the camsv, the seninf
source pad index is hardcoded, using a macro named
MTK_CAMSV_CIO_PAD_SRC. Not only is this confusing as the macro name
doesn't refer to the seninf, but it risks getting out of sync when
modifying the seninf driver.

Locate the source pad dynamically instead.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>