UPSTREAM_OR_FROMLIST: media: platform: mtk-isp: camsv: Don't propagate formats to subdevs
In a driver based on the media controller API, formats are propagated
across the pipeline by userspace. The video nodes must not perform any
format propagation inside the kernel, but must instead validate the
configuration of the pipeline when starting streaming.

Replace the format propagation with format validation at streamon time.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>