UPSTREAM_OR_FROMLIST: media: platform: mtk-isp: camsv: Use 1X16 YUV formats
While there is no documentation detailing whether the YUV formats are
transmitted on a 8- or 16-bit bus internally, a 16-bit bus is much more
probable. This matches the media bus code used for YUV formats on CSI-2,
simplifying handling of YUV sensors. Replace the 2X8 YUV formats with
the 1X16 variants.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>