UPSTREAM_OR_FROMLIST: media: platform: mtk-isp: seninf: dphy: De-duplicate register macros
The CSI1 and CSI2 instances are mostly identical, and very similar to
the CSI0 instance. Duplicating all register-related macros just makes
the code more complex. De-duplicate them, keeping macros specific to
CSI0 and CSI1/2 in the minority of cases where it is applicable.

This fixes a for CSI1 and CSI2 caused by the ANA18 and ANA1C registers
address being offset twice, once in the MIPI_RX_ANA18_CSI1A macro, and a
second time when adding the port-specific offset.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>