media: platform: mtk-isp: seninf: dphy: Don't configure B side in 2D1C mode
When operating in 4D1C mode, both sides of the PHY (CSIxA and CSIxB)
need to be configured. However, in 2D1C mode (only available for CSI0A
and CSI0B), the two sides are independent, and configuring the A side
must not modify the configuration of the B side. Worse, when configuring
CSI0B, the driver tries to add the B side offset to the register
addresses, and ends up configuring CSI1A.

Skip the configuration of the B side when operating in 2D1C mode.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>