UPSTREAM_OR_FROMLIST: media: platform: mtk-isp: seninf: Map inputs to DT ports
The relationship between device tree ports, CSI ports, SENINF and inputs
is messy. The CFG_CSI_PORT number identifies the CSI ports or sub-ports,
numbered from 0 to 4, but the port number is also used as a device tree
input port index, numbered from 0 to 3. The CSI0B sub-port is thus
unsupported.

The device tree bindings define four ports, with port 0 being either
CSI0 or CSI0A, depending on the mode in which the port is used. As CSI0
and CSI0A can't be used at the same time, and both map to SENINF1, this
design makes sense, and the driver has to be adapted.

In order to comply with the device tree bindings, map the port number in
the driver to the device tree port. CFG_CSI_PORT_0A is thus dropped, and
so are CFG_CSI_PORT_MAX_NUM and CFG_CSI_PORT_NONE as they're unused.

The only place where the port number is used in a non-compliant way with
the device tree is when configuring the mode and clock lane selector for
the CSI ports. The configuration of the CSI0 port now needs to take into
account both CFG_CSI_PORT_0 and CFG_CSI_PORT_0B, which conflicts with
the design of the mtk_seninf_set_csi_mipi() function that configures a
particular input. Fix this by moving the CSI ports configuration to a
new mtk_seninf_setup_phy() function that configures all CSI PHYs.

As a consequence, the device tree files need to be updated to use port@0
for CSI0A.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>