media: platform: mtk-isp: seninf: Decouple active_input from port number
The struct mtk_seninf active_input field stores the active port number
and is used to index in the inputs array. To refactor the
mtk_seninf_input structure to refer to a SENINF input channel, the two
concepts need to be decoupled. Do so by storing the port number and the
SENINF input index in the mtk_seninf_input structure, and use them
appropriately.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>