media: i2c: ap1302: Fix size calculation when setting format
The ap1302_find_best_size() determines the best size among the supported
values for the user requested size. To do so, it computes a score for
each possible size, but does it incorrectly by comparing supported sizes
with the full sensor resolution, not with the user requested size. Fix
it.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>