ANDROID: MTK: mt8183-pumpkin-ar0330: remove firmware node
With commit 030ddf2d6cac ("ANDROID: MTK: arm64: mt8183-pumpkin: remove
fstab node") we removed the fstab firmware node.

This is mandatory since early mount is different for boot header v2.

However, this node was not removed in mt8183-pumpkin-ar0330, resulting
in a boot loop whenever we attempt to boot with this dtbo.

Fix it by removing the node.

Fixes: 030ddf2d6cac ("ANDROID: MTK: arm64: mt8183-pumpkin: remove fstab node")
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>