media: platform: mtk-isp: camsv: Set device for camsv subdev
The camsv subdev's device is NULL in v4l2_device_register_subdev_nodes().
This leads to __video_register_device() falling back to the parent device
of associated v4l2_device struct, which often does not match the physical
device the subdevice is registered for.

For the camsv subdev case, it matches. But it's better to explicitly set
the device for each subdev as done in the seninf driver.

Signed-off-by: pnguyen <pnguyen@baylibre.com>