media: platform: mtk-isp: camsv: Fill buffers with timestamps
The V4L2 API requires the driver to report timestamps for buffers. Do
so, using the monotonic clock timestamp captured at end of frame for
simplicity.

Switching to start of frame as the timestamp source would give better
accuracy for applications, but V4L2 only support start of exposure and
end of frame. Record this issue in a comment for later development.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>