UPSTREAM_OR_FROMLIST: media: platform: mtk-isp: seninf: Propagate format from sink to source
The V4L2 subdev API requires subdev drivers to propagate formats from
sink to source. The SENINF can't perform any format conversion, so
propagation is easy. We however have to still allow setting formats on
the source pad for the time being, as the test pattern generator use
case would otherwise be broken. Switching to the experimental subdev
S_ROUTING API would be a good candidate to solve this issue.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>