media: platform: mtk-mdp3: Don't set media device dev field
The media device is normally a child of the physical device it relates
to. This cases issues with the MDP3, as creating a child device results
in the runtime PM API being used on the parent, which the MDP3 driver
doesn't handle correctly. The MDP3 driver then fails to communicate with
the SCP at runtime. Until this gets fixed, don't set the dev field of
the media device.

Set driver_name to avoid a crash in media_device_get_info() due to the
fallback path not being able to handle dev being NULL.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>