media: i2c: ap1302: Don't use devm_* API to acquire regulators
devm_regulator_bulk_get() ties the lifetime of the regulators with the
struct device passed to the function, which is also used to lookup the
regulators. The ap1302 driver uses the ap1302 struct device for this
purpose, which is incorrect as the regulators are specified in DT in the
sensors child nodes. Fixing this issue will require creating a device
for the sensors, to be used to lookup regulators, but the regulators
lifetime should still be bound to the ap1302. To support this, use
regulator_bulk_get() and manage regulators lifetime manually.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>