UPSTREAM_OR_FROMLIST: media: i2c: ap1302: Require all sensors to have the same model
The AP1302, when operating in dual sensor mode, require both sensors to
be identical. Enforce this limitation in the DT bindings and in the
driver. This allows storing a single instance of sensor info instead of
per-sensor information.

Update the firmware name scheme accordingly, as specifying the sensor
name twice in the firmware file name is useless.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>