BL_INTERNAL:HACK: gpu: drm: mediatek: reorder main and ext path on mt8167/mt8183
Re-order the main and ext path on MT8167/MT8183 to workaround some
CRTCs issues. encoder->possible_crtcs of DPI is not set
correctly when DSI is unused. As a workaround we change the order
of both path so that DPI is always possible_crtcs = BIT(0) in order for
this issue to never arise.

Signed-off-by: Fabien Parent <fparent@baylibre.com>