Input: edt-rpi-ft5x06: Add pumpkin rpi touchscreen support
This driver is a copy of edt-ft5x06.c modified to support the new LCD
module on RPi 7 touchscreen LCD
the last raspberry pi 7' touchscreen panel have not same touchscreen
controller and firmware version.
touch event was not return properly so we need to do polling to check
data and generate missing touch event where it's needed.

Bug: RITA-423
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>