ANDROID: MTK: mt8167-pumpkin-android: move dai-link-hdmi to overlay
On Android, we have the following device trees:
- mt8167-pumpkin.dts: main board device tree
- mt8167-pumpkin-android: default overlay, which uses HDMI
- mt8167-pumpkin-android-dsi: optional overlay, which replaces HDMI
usage with the RPi DSI panel. This overlay *disables* the &hdmi node.

Currrently, we *always* add the hdmi related dai-link, since it's
located in mt8167-pumpkin.dts.

When using mt8167-android-android-dsi, the sound-card probes in a loop,
returning -EPROBE_DEFER, because:
- we attempt to initialize the dai-link-hdmi
- dai-link-name waits for the &hdmi node to be up
- but &hdmi node is disabled

This is observed in the logs with the following error message:

[    2.089438] mt8516-soc-pumpkin sound: mt8516_pumpkin_dev_probe set_card_codec_info failed -517

Move the hdmi related sound nodes over to mt8167-pumpkin-android.dts so
that it only gets initialized when using HDMI.

Fixes: 84a338aca960 ("sound: soc: mediatek: mt8167: Dynamically add the hdmi audio codec")
Bug: RITA-420
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>