media: i2c: ap1302: Fix access to advanced registers
Advanced register addresses make use of bits 24, 25 and 30. This
overlaps with the size encoded in bits [26:24]. Fix it by moving the
size to bits [28:26] and updating the AP1302_REG_PAGE_MASK,
AP1302_REG_PAGE() and AP1302_REG_ADDR() macros accordingly.

Additionally, the advanced register space mapping window size is 4kB,
not 64kB. The AP1302_REG_PAGE_MASK must be updated accordingly to mask
out bits [15:12], but this results in a new issue: base registers in the
range 0x1000 to 0xffff will have a non-zero page value as computed by
AP1302_REG_PAGE(). The non-zero page value can't thus be used to check
if a register is a base or advanced register anymore. To fix this, use
bit 31 to mark advanced registers explicitly.

Signed-off-by: Stefan Hladnik <stefan.hladnik@gmail.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>