BL_INTERNAL:media: mtk-vcu: fixup IPI id for MT8183 VENC
The MT8183 venc is a non-OAL encoder, and uses the non-HYBRID IPI id, but this vcu
is used behind the VPU API, thus the wrong IPI is used to communicate with the vpud
daemon.

Add a simple fixup function to translate the VPU IPI to used to correct VCU IPI id
to communicate with the vpud daemon.

The other VPU IPI ids are correct so far, and doesn't break the MT8167 implementation.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>