UPSTREAM_OR_FROMLIST: media: i2c: ap1302: Decouple SIPM and AP1302 register encodings
Register addresses and sizes are encoded using the same scheme for the
AP1302 registers and the SIPM registers. The AP1302 register macros are
used to decode SIPM registers. This is a driver implementation
coincidence, not a hardware property. To prepare for a chance in the
AP1302 register encoding, decouple it from the SIPM register encoding.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>