ANDROID: MTK: arm64: mt8167-pumpkin: Add an audio i2s overlay for Android
Add an overlay in order to be able to use the I2S2 2 channels in and the
I2S1 8 channels out devices present on the 40 pins header.
In order to be able to fully use I2S1 at the same time as I2S2, the SD
card (sharing its pins with I2S1) must be disabled and the mux must be
switched to the 40pins header.

Signed-off-by: Nicolas <nbelin@baylibre.com>