sound: soc: mediatek: mt8167: Create 2 distinct I2S and HDMI devices
Create 2 distinct devices for HDMI and I2S 8ch devices even if they
are using the same fe and be DAIs. This brings more clarity regarding
the names and allows using the I2S 8ch without calling the hdmi codec
which is tied to the HDMI fe DAI. This was actually caussing issues
as the hdmi codec handles only 2ch when the i2s 8ch DAI goes up to 8ch.
The new device is put at the end of fe DAI list so the device indexes
are not going to be changed.

Signed-off-by: Nicolas Belin <nbelin@baylibre.com>