ANDROID: MTK: mtk_android_defconfig: move mt8167 modules to builtin
A lot of drivers used by the mt8167-pumpkin board are built as module
for yocto.

Since our Android is not configured to handle modalias properly, it's
simpler to have these built into the kernel.

This helps us avoiding manual 'insmod' and such.

Note that some drivers (such as mediatek-drm) are handled by our Android
OS as module, so these stay as modules.

Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>