BL_INTERNAL:media: mtk-vcodec-mt8167: correctly finish last encoding for OAL-vcodec
On MT8176, the codec is "hybrid" meaning the first encoding is delayed, the first
IPI call will be completed on the next call, thus it had a special STOP sequence
with a 0 sized packet to finish encoding.
Now we have a proper STOP/START commands for that, convert the code to use
the last source buffer indication to properly process the last encoding.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>