BL_INTERNAL:input: keyboard: mtk-kpd fix hardcoded row_shift value
The MATRIX_SCAN_CODE() macro take a 3rd argument: row_shift. As the
keycode are stored as a bitmap, the row are encoded as
(row << row_shift). In order to interpret correctly the rows keycode
we need to supply correctly the row_shift argument. This argument
is equal to the number of bit used to store the columns.

Replace the hardcoded row_shift and correctly compute that value
based on n_cols.

Signed-off-by: Fabien Parent <fparent@baylibre.com>