media:platform:mtk-isp: workaround for both camsv/camctl to build together
This workaround allow camsv and camctl to be built together in the same
image.
In the current architecture the probe() function is contained in the
'mtk_cam*_hw.c' files and call functions in common .c files that then
call back HW functions in 'mtk_cam*_hw.c' files.
The architecture shall be rethink. This workaround is for short term
releases.

Signed-off-by: FSylvestre <fsylvestre@baylibre.com>