[REVERT ME]: media: i2c: ov5645: Remove the full resolution mode
The OV5645 has the full resolution mode working at 15 fps while the
full HD mode and lower modes can go up to 30 fps.

Normally, the libcamera should choose the appropriate mode depending
on the requested resolution from user space. Unfortunately, it is not
the case in the current libcamera code.

So, awaiting for the changes in libcamera, we do this hack for the
time being to favor the framerate over the resolution.

This "fixes" the following Jira:

https://baylibre.atlassian.net/browse/MAN-39

Signed-off-by: Phi-Bang Nguyen <pnguyen@baylibre.com>