misc: mediatek: usb11: Drop incorrect __init annotation
The musbfsh_probe() and mt_usb11_probe() functions are marked as __init,
but they can be called after initialization. Drop their incorrect
annotation, as well as the wrong __init and __initdata annotations of
all functions and variables that they reference, directly or indirectly.

This issue was reported by the CONFIG_DEBUG_SECTION_MISMATCH kernel
configuration option.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>