REVERTME: ANDROID: MTK: Set bluetooth btmtk_sdio as built-in
The btmtk_sdio is linked to the mt7663/mt7668 combo wifi/bt chip.
We have two drivers using the same chip and the same sdio lines.
It creates restrictions on how both drivers should be probed and
 managed.

Probing the two drivers at the same time breaks both of the
 functionalities, so, as a temporary hack, set the btmtk_sdio
 driver to be built-in in order to artificially probe it earlier.

The proper fix may be to use a device_link() between both drivers.
But as of today, none of the drivers are using DT...

Signed-off-by: Guillaume Ranquet <granquet@baylibre.com>