BL_INTERNAL:input: mtk-kpd: Enable IRQ wakeup if necessary
If the devicetree has a property 'wakeup-source' the Mediatek keypad
driver is supposed to act as a possible component triggering a wakeup.

This is currently not happening as the interrupt is not handled as a
wakeup source.

The patch adds calls to enable_irq_wake and disable_irq_wake to define
these as IRQs that will wakeup the system from sleep states. As this IRQ
is always responsible for waking the system, enable_irq_wake is called
in probe() and disabled in remove(). The necessary data is added to
struct mtk_keypad and mtk_keypad is added as driver data to the device.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>