usb: xhci-mtk: Enable IRQ wakeup
The chip is already setup in the correct wakeup state, but the IRQ is
not used as a wakeup source right now as the IRQ is not marked for
wakeups.

This patch calls enable/disable_irq_wake() in usb_wakeup_ip_sleep_set()
to be able to wakeup the SoC when requested.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>