HACK: mt76: retry firmware loading when acquiring the semaphore fails
Due to the nature of this driver, it has to acquire a semaphore in order
to perform some of its operations as wifi and bluetooth functionnality
are shared on the same sdio bus.

Firmware loading can fail due to the impossibility to acquire the
semaphore (for reasons that are unknown to me yet...).

As of today, the firmware loading is done in a work queue and the error is not
propagated back... so the driver ends up being probed correctly with a
network device that you are unable to use.

	mt7663s mmc1:0001:1: Failed to get patch semaphore
	mt7663s mmc1:0001:1: mediatek/mt7663pr2h.bin not found, switching to mediatek/mt7663pr2h_rebb.bin
	mt7663s mmc1:0001:1: Failed to get patch semaphore
	mt7663s mmc1:0001:1: failed to load mediatek/mt7663pr2h_rebb.bin

As a workaround, sleep and retry as long as the semaphore cannot be acquired.

The proper fix would be a rewrite of these drivers to make it aware of its
companion on the same chip.

Signed-off-by: Guillaume Ranquet <granquet@baylibre.com>