WORKAROUND: ANDROID: MTK: Bluetooth: btmtksdio: don't touch data ram for MT7668u
When sending this command to a MT7668 chip with the public firmware from
[1], the chip seems to go into a wrong state:

[  231.742277] Bluetooth: hci0: Invalid bt type 0x80
[  231.743871] Bluetooth: hci0: Invalid bt type 0x80
[  231.744671] Bluetooth: hci0: Invalid bt type 0x80
[  231.746256] Bluetooth: hci0: Invalid bt type 0x80

Only send the "data RAM power on" for MT7663 chips.

Note: Commit a728697b74b7 ("Bluetooth: btmtksdio: fix up firmware download
sequence") is suspicious because there are no FW updates done on
mt7668pr2h.bin after this commit.
This makes me believe that it has not been retested on on MT7668.

This is marked as a work-around since:
- issue only occurs on Android (not on yocto)
- seems to be something performance related.

In any case, this fixes many CTS issues from CtsBluetoothTestCases.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/firmware/linux-firmware.git/tree/mediatek/mt7668pr2h.bin
Fixes: a728697b74b7 ("Bluetooth: btmtksdio: fix up firmware download sequence")
Bug: RITA-578
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>