drm/mediatek: don't break dither->dsi0 link on mt8183
The dither->dsi0 path is hardcoded to use register (MMSYS+0xf68) but
this register is actually at a different offset on mt8183, so we are
wrongly overriding the DISP_RDMA 0_RSZ_OUT _SEL_IN register on
MT8183 which break the DSI path.

This commit only update the dither->dsi0 path if a
dsi0_sel_in_dither attribute is set into the mmsys_reg_data
structure.

Signed-off-by: Fabien Parent <fparent@baylibre.com>