BL_INTERNAL: media: platform: mtk-isp: seninf: Split MUX and TOP MUX configuration
The MUX and TOP MUX (as well as the CAM MUX) are different entities,
despite having similar names. The TOP MUX and CAM MUX are global
crossbar switches, while the MUX is a data channel that includes a VC/DT
filter and a FIFO. Split their configuration in two different functions.

The MUX configuration function, mtk_seninf_mux_setup(), takes an
mtk_seninf_input pointer, while it should take a MUX object instead.
This will be fixed in a later change.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Pierre Moos <pmoos@baylibre.com>